refactor: 여행코스 페이지 테스트를 컴포넌트 테스트로 분리 #75
